split test files in header and cpp
